+++ api/src/main/java/org/pmiops/workbench/access/AccessSyncServiceImpl.java
@@ -14,13 +14,13 @@ import java.util.logging.Logger;
 import org.javers.common.collections.Lists;
 import org.pmiops.workbench.actionaudit.Agent;
 import org.pmiops.workbench.actionaudit.auditors.UserServiceAuditor;
+import org.pmiops.workbench.billing.FreeTierBillingService;
 import org.pmiops.workbench.config.WorkbenchConfig;
 import org.pmiops.workbench.db.dao.UserDao;
 import org.pmiops.workbench.db.model.DbAccessModule.DbAccessModuleName;
 import org.pmiops.workbench.db.model.DbAccessTier;
 import org.pmiops.workbench.db.model.DbUser;
 import org.pmiops.workbench.db.model.DbUserInitialCreditsExpiration;
-import org.pmiops.workbench.initialcredits.InitialCreditsExpirationService;
 import org.pmiops.workbench.institution.InstitutionService;
 import org.pmiops.workbench.model.Institution;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -36,7 +36,7 @@ public class AccessSyncServiceImpl implements AccessSyncService {
   private final AccessModuleService accessModuleService;
   private final InstitutionService institutionService;
   private final UserDao userDao;
-  private final InitialCreditsExpirationService initialCreditsExpirationServiceImpl;
+  private final FreeTierBillingService freeTierBillingService;
   private final UserServiceAuditor userServiceAuditor;
 
   @Autowired
@@ -46,14 +46,14 @@ public class AccessSyncServiceImpl implements AccessSyncService {
       AccessModuleService accessModuleService,
       InstitutionService institutionService,
       UserDao userDao,
-      InitialCreditsExpirationService initialCreditsExpirationServiceImpl,
+      FreeTierBillingService freeTierBillingService,
       UserServiceAuditor userServiceAuditor) {
     this.workbenchConfigProvider = workbenchConfigProvider;
     this.accessTierService = accessTierService;
     this.accessModuleService = accessModuleService;
     this.institutionService = institutionService;
     this.userDao = userDao;
-    this.initialCreditsExpirationServiceImpl = initialCreditsExpirationServiceImpl;
+    this.freeTierBillingService = freeTierBillingService;
     this.userServiceAuditor = userServiceAuditor;
   }
 
@@ -97,7 +97,7 @@ public class AccessSyncServiceImpl implements AccessSyncService {
       if (previousAccessTiers.isEmpty()
           && !newAccessTiers.isEmpty()
           && null == maybeCreditsExpiration) {
-        initialCreditsExpirationServiceImpl.createInitialCreditsExpiration(dbUser);
+        freeTierBillingService.createInitialCreditsExpiration(dbUser);
       }
     }
   }
