+++ api/src/main/java/org/pmiops/workbench/api/WorkspacesController.java
@@ -38,7 +38,6 @@ import org.pmiops.workbench.exceptions.ServerErrorException;
 import org.pmiops.workbench.exceptions.TooManyRequestsException;
 import org.pmiops.workbench.firecloud.FireCloudService;
 import org.pmiops.workbench.iam.IamService;
-import org.pmiops.workbench.initialcredits.InitialCreditsExpirationService;
 import org.pmiops.workbench.model.ArchivalStatus;
 import org.pmiops.workbench.model.CloneWorkspaceRequest;
 import org.pmiops.workbench.model.CloneWorkspaceResponse;
@@ -83,7 +82,6 @@ public class WorkspacesController implements WorkspacesApiDelegate {
   private final FireCloudService fireCloudService;
   private final FreeTierBillingService freeTierBillingService;
   private final IamService iamService;
-  private final InitialCreditsExpirationService initialCreditsExpirationService;
   private final Provider<DbUser> userProvider;
   private final Provider<WorkbenchConfig> workbenchConfigProvider;
   private final TaskQueueService taskQueueService;
@@ -104,7 +102,6 @@ public class WorkspacesController implements WorkspacesApiDelegate {
       FireCloudService fireCloudService,
       FreeTierBillingService freeTierBillingService,
       IamService iamService,
-      InitialCreditsExpirationService initialCreditsExpirationService,
       Provider<DbUser> userProvider,
       Provider<WorkbenchConfig> workbenchConfigProvider,
       TaskQueueService taskQueueService,
@@ -122,7 +119,6 @@ public class WorkspacesController implements WorkspacesApiDelegate {
     this.fireCloudService = fireCloudService;
     this.freeTierBillingService = freeTierBillingService;
     this.iamService = iamService;
-    this.initialCreditsExpirationService = initialCreditsExpirationService;
     this.taskQueueService = taskQueueService;
     this.userDao = userDao;
     this.userProvider = userProvider;
@@ -197,7 +193,7 @@ public class WorkspacesController implements WorkspacesApiDelegate {
       throw e;
     }
     final Workspace createdWorkspace =
-        workspaceMapper.toApiWorkspace(dbWorkspace, fcWorkspace, initialCreditsExpirationService);
+        workspaceMapper.toApiWorkspace(dbWorkspace, fcWorkspace, freeTierBillingService);
     workspaceAuditor.fireCreateAction(createdWorkspace, dbWorkspace.getWorkspaceId());
 
     if (dbWorkspace.isCDRAndWorkspaceTanagraEnabled()) {
@@ -278,7 +274,7 @@ public class WorkspacesController implements WorkspacesApiDelegate {
                             op,
                             workspaceDao,
                             fireCloudService,
-                            initialCreditsExpirationService,
+                            freeTierBillingService,
                             workspaceMapper)))
         .orElse(ResponseEntity.notFound().build());
   }
@@ -306,7 +302,7 @@ public class WorkspacesController implements WorkspacesApiDelegate {
               "processWorkspaceTask: begin processing operation %d by transitioning from %s to %s",
               operation.getId(),
               operation.getStatus().toString(),
-              DbWorkspaceOperationStatus.PROCESSING.toString()));
+              DbWorkspaceOperationStatus.PROCESSING));
       operation =
           workspaceOperationDao.save(operation.setStatus(DbWorkspaceOperationStatus.PROCESSING));
 
@@ -493,7 +489,7 @@ public class WorkspacesController implements WorkspacesApiDelegate {
     }
 
     final Workspace originalWorkspace =
-        workspaceMapper.toApiWorkspace(dbWorkspace, fcWorkspace, initialCreditsExpirationService);
+        workspaceMapper.toApiWorkspace(dbWorkspace, fcWorkspace, freeTierBillingService);
 
     int version = Etags.toVersion(workspace.getEtag());
     if (dbWorkspace.getVersion() != version) {
@@ -543,7 +539,7 @@ public class WorkspacesController implements WorkspacesApiDelegate {
     }
 
     final Workspace editedWorkspace =
-        workspaceMapper.toApiWorkspace(dbWorkspace, fcWorkspace, initialCreditsExpirationService);
+        workspaceMapper.toApiWorkspace(dbWorkspace, fcWorkspace, freeTierBillingService);
 
     workspaceAuditor.fireEditAction(
         originalWorkspace, editedWorkspace, dbWorkspace.getWorkspaceId());
@@ -653,7 +649,7 @@ public class WorkspacesController implements WorkspacesApiDelegate {
 
     dbWorkspace = workspaceDao.saveWithLastModified(dbWorkspace, user);
     final Workspace savedWorkspace =
-        workspaceMapper.toApiWorkspace(dbWorkspace, toFcWorkspace, initialCreditsExpirationService);
+        workspaceMapper.toApiWorkspace(dbWorkspace, toFcWorkspace, freeTierBillingService);
 
     workspaceAuditor.fireDuplicateAction(
         fromWorkspace.getWorkspaceId(), dbWorkspace.getWorkspaceId(), savedWorkspace);
@@ -840,7 +836,7 @@ public class WorkspacesController implements WorkspacesApiDelegate {
                     workspaceMapper.toApiRecentWorkspace(
                         dbWorkspacesById.get(userRecentWorkspace.getWorkspaceId()),
                         workspaceAccessLevelsById.get(userRecentWorkspace.getWorkspaceId()),
-                        initialCreditsExpirationService))
+                        freeTierBillingService))
             .collect(Collectors.toList());
     recentWorkspaceResponse.addAll(recentWorkspaces);
     return ResponseEntity.ok(recentWorkspaceResponse);
@@ -863,7 +859,7 @@ public class WorkspacesController implements WorkspacesApiDelegate {
     RecentWorkspaceResponse recentWorkspaceResponse = new RecentWorkspaceResponse();
     RecentWorkspace recentWorkspace =
         workspaceMapper.toApiRecentWorkspace(
-            dbWorkspace, workspaceAccessLevel, initialCreditsExpirationService);
+            dbWorkspace, workspaceAccessLevel, freeTierBillingService);
     recentWorkspaceResponse.add(recentWorkspace);
     return ResponseEntity.ok(recentWorkspaceResponse);
   }
