+++ api/src/main/java/org/pmiops/workbench/workspaces/WorkspaceOperationMapper.java
@@ -3,10 +3,10 @@ package org.pmiops.workbench.workspaces;
 import java.util.Optional;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
+import org.pmiops.workbench.billing.FreeTierBillingService;
 import org.pmiops.workbench.db.dao.WorkspaceDao;
 import org.pmiops.workbench.db.model.DbWorkspaceOperation;
 import org.pmiops.workbench.firecloud.FireCloudService;
-import org.pmiops.workbench.initialcredits.InitialCreditsExpirationService;
 import org.pmiops.workbench.model.Workspace;
 import org.pmiops.workbench.model.WorkspaceOperation;
 import org.pmiops.workbench.utils.mappers.MapStructConfig;
@@ -22,7 +22,7 @@ public interface WorkspaceOperationMapper {
       DbWorkspaceOperation source,
       WorkspaceDao workspaceDao,
       FireCloudService fireCloudService,
-      InitialCreditsExpirationService initialCreditsExpirationService,
+      FreeTierBillingService freeTierBillingService,
       WorkspaceMapper workspaceMapper) {
 
     WorkspaceOperation modelOperation = toModelWithoutWorkspace(source);
@@ -34,7 +34,7 @@ public interface WorkspaceOperationMapper {
                     workspaceId,
                     workspaceDao,
                     fireCloudService,
-                    initialCreditsExpirationService,
+                    freeTierBillingService,
                     workspaceMapper))
         .ifPresent(modelOperation::workspace);
 
@@ -57,7 +57,7 @@ public interface WorkspaceOperationMapper {
       long workspaceId,
       WorkspaceDao workspaceDao,
       FireCloudService fireCloudService,
-      InitialCreditsExpirationService initialCreditsExpirationService,
+      FreeTierBillingService freeTierBillingService,
       WorkspaceMapper workspaceMapper) {
     return workspaceDao
         .findActiveByWorkspaceId(workspaceId)
@@ -71,6 +71,6 @@ public interface WorkspaceOperationMapper {
                             workspaceMapper.toApiWorkspace(
                                 dbWorkspace,
                                 fcWorkspaceResponse.getWorkspace(),
-                                initialCreditsExpirationService)));
+                                freeTierBillingService)));
   }
 }
